fix: 添加卸载更新确认弹框，减少误操作 chore: 优化新增待办动画，及显示效果

+++ frontend/src/stores/global.ts
@@ -152,16 +152,28 @@ export const useGlobalStore = defineStore('global', {
             return this.update.updating;
         },
         uninstallUpdate() {
-            uninstallUpdate([
-                '/.vite',
-                '/assets',
-                '/index.html',
-                '/launch.html',
-                '/favicon.ico',
-            ]).then(() => {
-                console.log('[卸载更新成功] 返回启动页');
-                window.location.replace('/launch.html');
-            });
+            showConfirmDialog({
+                title: '提示',
+                message: '确定要卸载当前版本吗？',
+                confirmButtonText: '确定',
+                confirmButtonColor: '#f56c6c',
+                cancelButtonText: '取消',
+            })
+                .then(() => {
+                    this.update.text = '卸载中...';
+
+                    uninstallUpdate([
+                        '/.vite',
+                        '/assets',
+                        '/index.html',
+                        '/launch.html',
+                        '/favicon.ico',
+                    ]).then(() => {
+                        console.log('[卸载更新成功] 返回启动页');
+                        window.location.replace('/launch.html');
+                    });
+                })
+                .catch(() => {});
         },
     },
 });
